Refactor directive
